moved active notes
