+++ chord-game/src/routes/interval.svelte
@@ -11,7 +11,7 @@
 	 */
 	import { chordMatch, noteName } from '../lib/Chord.svelte';
 	import { chordGeneratorRandom } from '../lib/ChordSequence.svelte';
-	import ActiveNotes from '../lib/ActiveNotes.svelte';
+	import ActiveNotes from '../component/ActiveNotes.svelte';
 	import { onMount } from 'svelte';
 	import { addKeysToArray } from '../lib/Util.svelte';
 	import { tweened } from 'svelte/motion';
