+++ chord-game/src/routes/random.svelte
@@ -3,7 +3,7 @@
 	import { chordGeneratorRandom } from '../lib/ChordSequence.svelte';
 	import { getChordProgression, addKeysToChordSequence } from '../lib/ChordProression.svelte';
 	import { chordMatch, noteName } from '../lib/Chord.svelte';
-	import ActiveNotes from '../lib/ActiveNotes.svelte';
+	import ActiveNotes from '../component/ActiveNotes.svelte';
 	import { fade, fly } from 'svelte/transition';
 	import { flip } from 'svelte/animate';
     import { tweened } from 'svelte/motion';
